Implement domain.Storer and migrate to int64
- Updated invcache and invdb stores to implement the domain.Storer interface.
- Migrated all quantity fields (OnHand, Reserved, Incoming) to int64 for consistency and overflow prevention.
- Updated transaction handling in NewWithTx to return the interface type.
- Standardized cache serialization and DB mapping for the new data types.

+++ services/inventory/internal/stores/invdb/model.go
@@ -9,9 +9,9 @@ import (
 type inventoryDB struct {
 	LocationID string    `db:"location_id"`
 	SKU        string    `db:"sku"`
-	OnHand     int       `db:"on_hand"`
-	Reserved   int       `db:"reserved"`
-	Incoming   int       `db:"incoming"`
+	OnHand     int64     `db:"on_hand"`
+	Reserved   int64     `db:"reserved"`
+	Incoming   int64     `db:"incoming"`
 	UpdatedAt  time.Time `db:"updated_at"`
 }
